+++ Nos/Views/FollowsView.swift
@@ -15,12 +15,13 @@ struct FollowsView: View {
     @EnvironmentObject var router: Router
     
     var followed: Followed
-    let syncTimer = SyncTimer()
-
-    @State private var authorsToSync: [Author] = []
-
-    func author(id: String) -> Author {
-        try! Author.findOrCreate(by: id, context: viewContext)
+    
+    @State private var subscriptionId: String = ""
+    
+    func refreshFollows() {
+        let keys = followed.compactMap { $0.destination?.hexadecimalPublicKey }
+        let filter = Filter(authorKeys: keys, kinds: [.metaData, .contactList], limit: 100)
+        subscriptionId = relayService.requestEventsFromAll(filter: filter)
     }
     
     var body: some View {
@@ -28,14 +29,8 @@ struct FollowsView: View {
             LazyVStack {
                 ForEach(followed) { follow in
                     VStack {
-                        FollowCard(author: follow.destination!)
-                    }
-                    .onAppear {
-                        // Error scenario: we have an event in core data without an author
-                        let author = follow.destination!
-                        if !author.isPopulated {
-                            print("Need to sync author: \(author.hexadecimalPublicKey ?? "")")
-                            authorsToSync.append(author)
+                        if let author = follow.destination {
+                            FollowCard(author: author)
                         }
                     }
                     Spacer()
@@ -46,14 +41,12 @@ struct FollowsView: View {
         .onAppear {
             router.navigationTitle = "Follows"
         }
-        .onReceive(syncTimer.currentTimePublisher) { _ in
-            if !authorsToSync.isEmpty {
-                print("Syncing \(authorsToSync.count) authors")
-                let authorKeys = authorsToSync.map({ $0.hexadecimalPublicKey! })
-                let filter = Filter(authorKeys: authorKeys, kinds: [.metaData], limit: authorsToSync.count)
-                relayService.requestEventsFromAll(filter: filter)
-                authorsToSync.removeAll()
-            }
+        .task {
+            refreshFollows()
+        }
+        .onDisappear {
+            relayService.sendCloseToAll(subscriptions: [subscriptionId])
+            subscriptionId = ""
         }
     }
 }